fix: widen u32 overflow sites to u64 in posting/position/sparse formats and TermInfo
Breaking format changes (no backward compat):

- BlockPostingList: footer data_len u32→u64, skip entry offset u32→u64
  (footer 16B→20B, skip entries 16B→20B)
- PositionPostingList: footer data_len u32→u64, skip entry offset u32→u64
  (footer 12B→16B, skip entries 12B→16B)
- TermInfo::External: posting_len u32→u64, position_len u32→u64
  (vint-encoded, wire format auto-adjusts)
- SparseSkipEntry: offset u32→u64 (entry 20B→24B)

Prevents silent data corruption when a single term's posting/position
data exceeds 4GB during segment merge.

+++ hermes-core/src/query/scoring.rs
@@ -977,7 +977,7 @@ struct LazyTermCursor {
     /// Number of skip entries (blocks) for this dimension
     skip_count: usize,
     /// Base byte offset for block data (pre-resolved, avoids dim_id lookup per load)
-    block_data_offset: u32,
+    block_data_offset: u64,
     /// Current block index (0-based relative to this dimension's blocks)
     block_idx: usize,
     /// Decoded block data (loaded on demand, reused across seeks)
@@ -997,7 +997,7 @@ impl LazyTermCursor {
         skip_start: usize,
         skip_count: usize,
         global_max_weight: f32,
-        block_data_offset: u32,
+        block_data_offset: u64,
     ) -> Self {
         let exhausted = skip_count == 0;
         let abs_qw = query_weight.abs();
